refactor : LottoTicketsService 삭제하고 로또 티켓들을 생성하는 책임을 LottoTickets 로 이동
